CHROMIUM: config: Enable NETFILTER_XT_MATCH_CGROUP
This patch enables NETFILTER_XT_MATCH_CGROUP (which selects
CGROUP_NET_CLASSID) in kernel config: enables net_cls in cgroup and its
support in netfilter. With this patch, we can put a process into a
net_cls cgroup and as a result, all the locally generated packets with
this process will be tagged with a given net classid, and then we can
match packets by this classid using iptables, for routing, traffic
metering, and other purposes.

Note that this option is added into the kernel from v3.14.

BUG=b:167479541
TEST=On atlas, able to mount net_cls cgroup and create iptables rules
  matching cgroup.

Signed-off-by: Jie Jiang <jiejiang@chromium.org>
Change-Id: I8756e435a6b2e274955b4989123eba14b980ee2f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2387983
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>