UPSTREAM: media: staging: rkisp1: params: upon stream stop, iterate a local list to return the buffers
The code in '.stop_streaming' callback releases and acquire the lock
at each iteration when returning the buffers.
Holding the lock disables interrupts so it should be minimized.
To make the code cleaner and still minimize holding the lock,
the buffer list is first moved to a local list and
then can be iterated without the lock.

Signed-off-by: Dafna Hirschfeld <dafna.hirschfeld@collabora.com>
Acked-by: Helen Koike <helen.koike@collabora.com>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>
Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>
(cherry picked from commit b1b2715e8d616bab68ed4e22b6617601105f2947)

BUG=b:168991270
TEST=Camera works on Scarlet
TEST=CtsCameraTestCases passes on Scarlet

Signed-off-by: Ricardo Ribalda <ribalda@chromium.org>
Change-Id: Ia522c9cfd007e94a6924fee6bc2f6a9605a60636
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2517832
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Tomasz Figa <tfiga@chromium.org>