UPSTREAM: media: staging: rkisp1: isp: add a warning and debugfs var for irq delay
The signal RKISP1_CIF_ISP_FRAME is set when the ISP completes
outputting the frame to the next block in the pipeline.
In order to keep buffer synchronization we assume that the
RKISP1_CIF_ISP_V_START signal never arrives together with the
RKISP1_CIF_ISP_FRAME signal.
In case those signals arrive together then the code is not able to
tell if the RKISP1_CIF_ISP_FRAME signal relates to the frame of
the current v-start or the previous. This patch adds a WARN_ONCE
and a debugfs var to catch it.

Signed-off-by: Dafna Hirschfeld <dafna.hirschfeld@collabora.com>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>
Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>
(cherry picked from commit 8524fa22fd2f49ca39834fe93c952ddac1c43750)

BUG=b:168991270
TEST=Camera works on Scarlet
TEST=CtsCameraTestCases passes on Scarlet

Signed-off-by: Ricardo Ribalda <ribalda@chromium.org>
Change-Id: Id1fa0271e6df2a8bb317ec3d432223b37e279750
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2517838
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Tomasz Figa <tfiga@chromium.org>