CHROMIUM: iwl7000: mac80211: ignore country element TX power on 6 GHz
Updates to the 802.11ax draft are coming that deprecate the
country element in favour of the transmit power envelope
element, and make the maximum transmit power level field in
the triplets reserved, so if we parse them we'd use 0 dBm
transmit power.

Follow suit and completely ignore the element on 6 GHz for
purposes of determining TX power.

Change-Id: Icb6e52af586edcc74f1f0360e8f6fc9ef2bfe8f5
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 5b53fa5881034bb0ea9b7175db46717c18344228