CHROMIUM: iwl7000: Revert "iwlwifi: take into acount the access bus time"
The commit below relied on jiffies to stop spinning when
polling for a bit to be set on a register on the device.
The problem is that we call iwl_poll_bit from
iwl_pcie_tx_stop_fh with interrupt disabled which means
that jiffies can't be incremented (since they are
incremented by a interrupt on the CPU). Bottom line: we
were spinning there with interrupts disabled until the
bit we're polling finally gets set.

This wouldn't be problem if the bit we're polling in the
register would be set quickly, but on old devices, there
is a flow in which the device is completely dead and the
bit would never get set, leading to a complete stall of
the system.

This reverts commit db17c1838e9f3e66d8cba63bbba1950725961f04.

Change-Id: I9b36cb33d2101955837b0c0e10b1fb878b6eeea7
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
(cherry picked from commit 80dabede8f8e9ba5bf664f8c6ee2f2c01307cb33)
iwl7000-tree: 27135709fdd860c614acfb4e6e7a6b9e21bd5658