CHROMIUM: iwl7000: mvm: implement approved list for the PPAG feature
We should only allow PPAG to be enabled by OEMs that are in the
approved list.  In order to do this, we need to compare the system
vendor string retrieved from SMBIOS to a list maintained in the
driver.  If the vendor string is not in the list, we don't allow PPAG
to be used.

For now the list is empty, but entries will be added to it
individually, in subsequent patches.

Change-Id: I4e4cf61d8fa6ff91d9b0cab2b1ec9ede4be346f5
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: 855fe04303d3f34ecd154fdf30c0c5a39e3c4490