CHROMIUM: iwl7000: mvm: fix the type we use in the PPAG table validity checks
The value we receive from ACPI is a long long unsigned integer but the
values should be treated as signed char.  When comparing the received
value with ACPI_PPAG_MIN_LB/HB, we were doing an unsigned comparison,
so the negative value would actually be treated as a very high number.

To solve this issue, assign the value to our table of s8's before
making the comparison, so the value is already converted when we do
so.

Change-Id: If77fd9c61a96aa7ef2ac96d935b7efd7df502399
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: fe4ec75bb74d8efe1c722e6c517066d51994d180