BACKPORT: media: uapi: add MEDIA_BUS_FMT_METADATA_FIXED media bus format.
MEDIA_BUS_FMT_METADATA_FIXED should be used when
the same driver handles both sides of the link and
the bus format is a fixed metadata format that is
not configurable from userspace.
The width and height will be set to 0 for this format.

Signed-off-by: Dafna Hirschfeld <dafna.hirschfeld@collabora.com>
Acked-by: Helen Koike <helen.koike@collabora.com>
Acked-by: Sakari Ailus <sakari.ailus@linux.intel.com>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>
Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>

(cherry picked from commit 6ad253cc3436269fc6bcff03d704c672f368da0a)

Conflicts:
   Documentation/userspace-api/media/v4l/subdev-formats.rst

BUG=b:168991270
TEST=Camera works on Scarlet
TEST=CtsCameraTestCases passes on Scarlet

Signed-off-by: Ricardo Ribalda <ribalda@chromium.org>
Change-Id: Ie01582dc3d586f1dd9989752371612650b30284f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2517851
Reviewed-by: Tomasz Figa <tfiga@chromium.org>