CHROMIUM: iwlwifi: only support GeoSAR v0 tables
Some existing GeoSAR tables incorrectly set the version field to 1
instead of 0 to mean that the contents of the table are invalid and
should be ignored.
However, recent versions of iwlwifi have started handling version 1 as
valid, resulting in incorrect GeoSAR tables being applied on those
devices.
Until we have audited and fixed those broken GeoSAR tables, keep
supporting only v0 tables.

BUG=b:172034758
TEST=load a v1 GeoSAR table that sets TX power to 0 on a device, observe
that the driver rejects the table and does not send the broken SAR table
to the firmware.

Change-Id: I029d2a2226f62d6201a3eda1bdf9745fa2e9f0ae
Signed-off-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2513541
Reviewed-by: Brian Norris <briannorris@chromium.org>