CHROMIUM: HACK: media: rkisp1: vb2: Fill data_offset of vb2_plane by that of v4l2_plane
With V4L2 MEMORY DMABUF for Multiplane API, it is necessary
to set a file descriptor on each plane. The way to pass an
offset with a DMA-buf is not defined in V4L2 specification.
Some drivers need to abuse the data_offset field for the
purpose. This patch fills the data_offset field of vb2_plane
by the same field of v4l2_plane, so drivers may utilize it
when needed.

BUG=b:168991270
TEST=Camera works on Scarlet
TEST=CtsCameraTestCases passes on Scarlet

Suggested-by: Helen Koike <helen.koike@collabora.com>
Signed-off-by: Ricardo Ribalda <ribalda@chromium.org>
Change-Id: Ib0dbee5941eb6bd65facc19f46bbba9e1e98335f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2526081
Reviewed-by: Tomasz Figa <tfiga@chromium.org>