CHROMIUM: iwl7000: mvm: register to mac80211 last
All the actions that were taken after the registration can
be taken before the registration to mac80211.
This will help to defer part of the op_mode_mvm_start
function to a later stage in case the device is not
available.

Change-Id: Ia928ca34d25a73d959a345ffbe4f1217c3f17394
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
iwl7000-tree: 1872537396a943f047cbde1abce647a0db91e829