CHROMIUM: Bluetooth: Remove hci_req_le_suspend_config
Since suspend task bits are set directly when configuring event filter
and passive scanning, there is no need for a separate function to do so.
As a result, both the suspend disable and suspend enable bit can be set
at the same time so the suspend_req_complete function should clear both.

DO NOT UPSTREAM THIS PATCH as-is. The upstream code currently doesn't
match our repo and this patch is missing pieces necessary for
upstreaming.

BUG=b:174184979, b:173799663
TEST=Built and tested with bluetooth_AdapterSRHealth

Signed-off-by: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>
Change-Id: Icb1766bc30b91c7cf1d4d5043340f9a604eaf652
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2576125
Reviewed-by: Daniel Winkler <danielwinkler@google.com>