FROMLIST: media: rkisp1: rsz: crash fix when setting src format
When setting the source media bus code in the resizer,
we first check that the current media bus code in the
source is yuv encoded format. This is done by
retrieving the data from the formats list of the isp
entity. This cause a crash when the media bus code on the
source is YUYV8_1_5X8 which is not supported by the isp
entity. Instead we should test the sink format of the resizer
which is guaranteed to be supported by the isp entity.

Fixes: 251b6eebb6c49 ("media: staging: rkisp1: rsz: Add support to more YUV encoded mbus codes on src pad")
Signed-off-by: Dafna Hirschfeld <dafna.hirschfeld@collabora.com>
Acked-by: Helen Koike <helen.koike@collabora.com>
Tested-by: Sebastian Fricke <sebastian.fricke.linux@gmail.com>
(am from http://patchwork.linuxtv.org/patch/69910/)
(also found at https://lore.kernel.org/r/20201209161510.31399-1-dafna.hirschfeld@collabora.com)

BUG=b:168991270
TEST=Camera works on Scarlet
TEST=CtsCameraTestCases passes on Scarlet

Signed-off-by: Ricardo Ribalda <ribalda@chromium.org>
Change-Id: I747b0d775a8e7b2c5634b7bad78f09736f322ae4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2585084
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Tomasz Figa <tfiga@chromium.org>