CHROMIUM: iwl7000: xvt: Check if pkts were txed before calling reclaim
When the fw sends txpath_flush notification, driver has to check if the
packet is transmitted after last flush. If the packet was really
transmitted then read_before and read_after will be different.
So driver has to call get_tx_meta_data and reclaim function only if
read_before and read_after are different.

In iwl_xvt_rx_get_tx_meta_data func, it is not necessary to check
queue_data.allocated_queue as get_tx_meta_data func is called only by tx
callback. So in this commit, we have removed checking of
queue_data.allocated_queue in get_tx_meta_data func.

In one of the pyfi test cases, we are sending infinite frames and then
immediately we are suspending the queue, queue contents are not
transmitted and queue becomes full. In this case we get a warning in
get_tx_meta_data func. So driver has to check if the queue is full in
tx_path_flush func. If the queue is full, then don't call get_tx_meta_data
func.

Change-Id: I99d770368c25d67cc091bf8c5d989d55987329e0
Signed-off-by: Abhishek Naik <abhishek.naik@intel.com>
iwl7000-tree: d19f9ed5fb97674be6df16dbb26fc14940619783