CHROMIUM: iwl7000: mvm: don't disconnect immediately if we don't hear beacons after CSA
When we switch channel, we may miss a few beacons on the
new channel. Don't disconnect if the time event for the
switch ends before we hear the beacons.

Note that this is relevant only for old devices that still
use the TIME_EVENT firmware API for channel switch.

The check that we hear a beacon before the time event
ends was meant to be used for the association time event
and not for the channel switch time event.

Change-Id: I37a161ffdfb099a10080fbdc3b70a4deb76952e2
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
iwl7000-tree: f2e2a82c2eecc95b6d80bad8a8a9184bd5591448