CHROMIUM: iwl7000: mvm: don't change band on bound PHY contexts
When we have a P2P Device active, we attempt to only change the
PHY context it uses when we get a new remain-on-channel, if the
P2P Device is the only user of the PHY context.

This is fine if we're switching within a band, but if we're
switching bands then the switch implies a removal and re-add
of the PHY context, which isn't permitted by the firmware while
it's bound to an interface.

Fix the code to skip the unbind/release/... cycle only if the
band doesn't change (or we have old devices that can switch the
band on the fly as well.)

Change-Id: I713b9d109957df7e7d9ed0861d5377ce3f8fccd3
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: a96fe7f66cad3f73b8c9f80d0ac4998dc12cf346